requirements/base.txt
# Python 3.11.9

Django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==1.0.0
coverage==6.5.0 